Change buying decision
Based on expermintation, stocks that are in a free fall always do bad
(Sears). Stocks that are going up always seem to do well but not as
well as buy and hold. We still can make money though on the BACs of
the world: always solid but not on a bubble spike like AAPL or NVDA.
This change helps protect us from not great stocks and tries to only buy
stocks when we "know" they will go up the next day